Cria e utiliza a extensao @Include
