+++ src/main/java/br/com/javamagazine/extension/RegionExpressionInterpreter.java
@@ -1,0 +1,19 @@
+package br.com.javamagazine.extension;
+
+import br.com.javamagazine.region.Region;
+import br.com.javamagazine.region.ServerWebServiceClient;
+
+public class RegionExpressionInterpreter implements ExpressionInterpreter<Region, Boolean> { 
+
+  @Override
+  public Boolean interpret(Region expression) {
+    
+    ServerWebServiceClient swsClient = new ServerWebServiceClient();
+        
+    // recupera a regiao através da api fornecida pelo servidor na nuvem.
+    Region bucketLocation = swsClient.getBucketLocation();
+    return bucketLocation.equals(expression);
+    
+  }
+
+}
