--- src/main/java/br/com/javamagazine/pagamento/Pagamento.java
@@ -1,19 +1,0 @@
-package br.com.javamagazine.pagamento;
-
-import java.math.BigDecimal;
-
-public interface Pagamento {
-
-	/**
-	 * Consulta valor de imposto para um pagamento.
-	 * @param valorPagamento
-	 * @return
-	 */
-	BigDecimal calcularTaxaImposto(BigDecimal valorPagamento);
-	
-	/**
-	 * Efetua um pagamento
-	 */
-	void efetuarPagamento(BigDecimal valorPagamento);
-	
-}
