--- src/main/java/br/com/javamagazine/pagamento/PagamentoVenezuela.java
@@ -1,23 +1,0 @@
-package br.com.javamagazine.pagamento;
-
-import java.math.BigDecimal;
-
-import br.com.javamagazine.extension.CountryExpressionInterpreter;
-import br.com.javamagazine.extension.Include;
-
-@Include(onExpression = "VE", interpretedBy = CountryExpressionInterpreter.class)
-public class PagamentoVenezuela implements Pagamento {
-
-	@Override
-	public BigDecimal calcularTaxaImposto(BigDecimal valorPagamento) {
-		return new BigDecimal("2.75");
-	}
-
-	@Override
-	public void efetuarPagamento(BigDecimal valorPagamento) {
-		BigDecimal taxaImposto = calcularTaxaImposto(valorPagamento);
-		BigDecimal pagamento = valorPagamento.multiply(taxaImposto).add(valorPagamento);
-		System.out.println(pagamento.toPlainString());
-	}
-
-}
