+++ src/main/resources/META-INF/services/javax.enterprise.inject.spi.Extension
@@ -1,1 +1,1 @@
-br.com.javamagazine.extension.CountryExtension
+br.com.javamagazine.extension.RegionExtension
