+++ src/main/webapp/WEB-INF/beans.xml
@@ -5,8 +5,4 @@
 							http://xmlns.jcp.org/xml/ns/javaee/beans_1_1.xsd"
 		version="1.1" bean-discovery-mode="all">
 
-  <alternatives>
-    <class>br.com.javamagazine.produtos.model.CalculadoraDeImpostosSA</class>
-  </alternatives>
-  
 </beans>
